ufdub: Drop AnimeSourceFactory, ConfigurableAnimeSource
